Batch 1 of clean-ups, autoboxing, deprecation swaps

+++ com.openexchange.configuration.clt/src/com/openexchange/configuration/clt/ExtractJULModificationsCLT.java
@@ -77,7 +77,7 @@ public class ExtractJULModificationsCLT extends AbstractCLI<Integer, Void> {
         if (retVal == null) {
             retVal = Integer.valueOf(1);
         }
-        System.exit(retVal);
+        System.exit(retVal.intValue());
     }
 
     /**
